plugins.py now is a library
