+++ lib/plugin_parser/__init__.py
@@ -1,0 +1,9 @@
+#!/usr/bin/python
+
+from .parser import Parser
+from .matcher import Matcher
+from .extractor import Extractor
+from .test import Test
+from .plugin import Plugin
+
+pass
